[PUBDEV-4274] Integrate XGBoost with H2O core as a separated extension module
Modularize XGBoost backend:
  - XGBoost implementation separated into two modules
    - genmodel support with native libraries and xgboost Java API
    - xgboost rest api and model builder
  - Add ability to load MOJO Reader by name via SPI.
  - MojoReader has a new method getModelName() to return model identifier.
  - Single node tests enabled for xgboost module
  - Proper bundling of h2o-genmodel-ext-xgboost
    - Includes: upgrade of shadow plugin from 1.2.4 to 2.0.0
  - Native libraries for linux_64 cpu/gpu, osx cpu (no OMP support)
  - Adding xgboost extension into assembly process of H2O
    - Includes:
      - Default H2O assembly contains xgboost by default, can be control via component api (-PwithXgboost=false)
      - h2o-genmodel contains xgboost genmodel if h2o.jar contains xgboost.
  - [HOTFIX] Various hotfixes in test infrastructure
      - Change path of h2o-genmodel.jar to a new location.
      - [HOTFIX] GenModel assembly was not properly plug into build graph
      - Forgot to regenerate R scripts
      - [FIX] Replacing Linux_64 xgboost4j.so
      - The lib does not contain statically linked CUDA library anymore but expects it give via LD_LIBRARY_PATH
      - [TEST] Test updated to check if test environment is compatible with XGboost
      - Minor infrastructure changes
        - assertion should be check after setup
        - test rest api test was missing xgboost as expected algo
