+++ build.gradle
@@ -12,7 +12,7 @@ buildscript {
     //noinspection GroovyAssignabilityCheck
     dependencies {
         classpath 'org.ow2.asm:asm:5.1'
-        classpath 'com.github.jengelman.gradle.plugins:shadow:1.2.4'
+        classpath 'com.github.jengelman.gradle.plugins:shadow:2.0.0'
         classpath 'org.gradle.api.plugins:gradle-nexus-plugin:0.7.1'
         classpath 'com.github.townsfolk:gradle-release:1.2'
         classpath 'com.adaptc.gradle:nexus-workflow:0.6'
@@ -69,7 +69,9 @@ ext {
       project(':h2o-orc-parser'),
       project(':h2o-parquet-parser'),
       project(':h2o-jaas-pam'),
-      project(':h2o-automl')
+      project(':h2o-automl'),
+      project(':h2o-genmodel-ext-xgboost'),
+      project(':h2o-ext-xgboost')
     ]
 
     javaProjects = [
@@ -87,7 +89,9 @@ ext {
       project(':h2o-orc-parser'),
       project(':h2o-parquet-parser'),
       project(':h2o-jaas-pam'),
-      project(':h2o-automl')
+      project(':h2o-automl'),
+      project(':h2o-genmodel-ext-xgboost'),
+      project(':h2o-ext-xgboost')
     ]
 
     scalaProjects = [
@@ -140,7 +144,9 @@ ext {
     // They can be enabled by passing parameter "-Pwith{componentName}=true"
     // For example, `-PwithVisDataServer=true -PvisDataServerVersion=3.14.27`
     optionalComponents = [
-            [name: "visDataServer", version : "1.0.0-SNAPSHOT", enabled: false ]
+            // NAME | VERSION | IS ENABLED? | HAS GENMODEL PART ?
+            [name: "visDataServer", version : "1.0.0-SNAPSHOT", enabled: false, genmodel: false ],
+            [name: "xgboost", version: "$version", enabled: true, genmodel: true ]
     ]
 
     //
