+++ gradle/components/xgboost.gradle
@@ -1,0 +1,4 @@
+dependencies {
+    compile project(":h2o-genmodel-ext-xgboost")
+    compile project(":h2o-ext-xgboost")
+}
