+++ h2o-extensions/xgboost/src/main/resources/META-INF/services/water.api.RestApiExtension
@@ -1,0 +1,1 @@
+hex.api.xgboost.RegisterRestApi
