+++ h2o-genmodel-extensions/xgboost/src/main/resources/META-INF/services/hex.genmodel.ModelMojoReader
@@ -1,0 +1,1 @@
+hex.genmodel.algos.xgboost.XGBoostMojoReader
