+++ h2o-genmodel/build.gradle
@@ -1,5 +1,3 @@
-apply plugin: 'com.github.johnrengelman.shadow'
-
 //
 // H2O Generated Model Support Module
 //
@@ -29,10 +27,4 @@ dependencies {
   compile "com.google.code.gson:gson:2.6.2"
   compile "com.google.protobuf.nano:protobuf-javanano:3.1.0"
   compile "ai.h2o:deepwater-backend-api:1.0.3"
-  if (file("${rootDir}/lib/xgboost4j-0.7-jar-with-dependencies.jar").exists()) {
-      compile files("${rootDir}/lib/xgboost4j-0.7-jar-with-dependencies.jar")
-  }
 }
-
-build.dependsOn shadowJar
-jar.dependsOn shadowJar
