+++ h2o-genmodel/src/main/java/hex/genmodel/ModelMojoFactory.java
@@ -7,15 +7,40 @@ import hex.genmodel.algos.glm.GlmMojoReader;
 import hex.genmodel.algos.glrm.GlrmMojoReader;
 import hex.genmodel.algos.kmeans.KMeansMojoReader;
 import hex.genmodel.algos.word2vec.Word2VecMojoReader;
-
-import hex.genmodel.algos.xgboost.XGBoostMojoReader;
+import java.util.ServiceLoader;
 
 /**
  * Factory class for instantiating specific MojoGenmodel classes based on the algo name.
  */
 public class ModelMojoFactory {
 
-  public static ModelMojoReader getMojoReader(String algo) {
+  public final static ModelMojoFactory INSTANCE = new ModelMojoFactory();
+
+  /** Service loader for model mojo readers.
+   *
+   * Based on JavaDoc of SPI: "Instances of this class are not safe for use by multiple concurrent threads." - all usages of the loader
+   * are protected by synchronized block.
+   */
+  private final ServiceLoader<ModelMojoReader> loader;
+
+  private ModelMojoFactory() {
+    loader = ServiceLoader.load(ModelMojoReader.class);
+  }
+
+  private ModelMojoReader loadMojoReader(String algo) {
+    assert algo != null : "Name of algorithm should be != null!";
+    synchronized (loader) {
+      loader.reload();
+      for (ModelMojoReader mrb : loader) {
+        if (algo.equals(mrb.getModelName())) {
+          return mrb;
+        }
+      }
+    }
+    return null;
+  }
+
+  public ModelMojoReader getMojoReader(String algo) {
     if (algo == null)
       throw new IllegalArgumentException("Algorithm not specified.");
 
@@ -29,9 +54,6 @@ public class ModelMojoFactory {
 
       case "Deep Water":
         return new DeepwaterMojoReader();
-
-      case "XGBoost":
-        return new XGBoostMojoReader();
 
       case "Generalized Low Rank Modeling":
       case "Generalized Low Rank Model":
@@ -48,8 +70,13 @@ public class ModelMojoFactory {
         return new KMeansMojoReader();
 
       default:
-        throw new IllegalStateException("Unsupported MOJO algorithm: " + algo);
+        // Try to load MOJO reader via service
+        ModelMojoReader mmr = loadMojoReader(algo);
+        if (mmr != null) {
+          return mmr;
+        } else {
+          throw new IllegalStateException("Unsupported MOJO algorithm: " + algo);
+        }
     }
   }
-
 }
