+++ h2o-genmodel/src/main/java/hex/genmodel/ModelMojoReader.java
@@ -27,7 +27,7 @@ public abstract class ModelMojoReader<M extends MojoModel> {
     if (! info.containsKey("algorithm"))
       throw new IllegalStateException("Unable to find information about the model's algorithm.");
     String algo = String.valueOf(info.get("algorithm"));
-    ModelMojoReader mmr = ModelMojoFactory.getMojoReader(algo);
+    ModelMojoReader mmr = ModelMojoFactory.INSTANCE.getMojoReader(algo);
     mmr._lkv = info;
     mmr._reader = reader;
     try {
@@ -39,6 +39,7 @@ public abstract class ModelMojoReader<M extends MojoModel> {
     return mmr._model;
   }
 
+  public abstract String getModelName();
 
   //--------------------------------------------------------------------------------------------------------------------
   // Inheritance interface: ModelMojoWriter subclasses are expected to override these methods to provide custom behavior
