+++ h2o-genmodel/src/main/java/hex/genmodel/algos/kmeans/KMeansMojoReader.java
@@ -7,6 +7,11 @@ import java.io.IOException;
 public class KMeansMojoReader extends ModelMojoReader<KMeansMojoModel> {
 
   @Override
+  public String getModelName() {
+    return "K-means";
+  }
+
+  @Override
   protected void readModelData() throws IOException {
     _model._standardize = readkv("standardize");
     if (_model._standardize) {
