+++ h2o-web/build.gradle
@@ -13,7 +13,7 @@ configurations {
 
 dependencies {
     // Dependency on shared model support which is served by web backend as a file
-    publishedJars project(path: ':h2o-genmodel', configuration: 'shadow')
+    publishedJars project(path: ':h2o-assemblies:genmodel', configuration: 'shadow')
 }
 
 static def getOsSpecificCommandLine(args) {
@@ -112,7 +112,7 @@ task copyFlowToWebRoot(type: Copy) {
 
 task copyGenModelJarToWebRoot(type: Copy) {
   from {
-      configurations.publishedJars.filter { it.name.startsWith("h2o-genmodel") }
+      configurations.publishedJars.filter { it.name.contains("genmodel") }
   }
   into "src/main/resources/www/${h2oRESTApiVersion}/"
   rename { filename -> "h2o-genmodel.jar"}
@@ -124,14 +124,6 @@ task copyDeepWaterJarToWebRoot(type: Copy) {
   }
   into "src/main/resources/www/${h2oRESTApiVersion}/"
 }
-
-task copyXGBoostJarToWebRoot(type: Copy) {
-  from {
-      file("${rootDir}/lib/xgboost4j-0.7-jar-with-dependencies.jar")
-  }
-  into "src/main/resources/www/${h2oRESTApiVersion}/"
-}
-
 
 task deleteBowerModules(type: Delete) {
   delete 'lib'
@@ -151,10 +143,6 @@ task deleteGenModelJarFromWebRoot(type: Delete) {
 
 task deleteDeepWaterJarFromWebRoot(type: Delete) {
     delete "src/main/resources/www/${h2oRESTApiVersion}/deepwater-all.jar"
-}
-
-task deleteXGBoostJarFromWebRoot(type: Delete) {
-    delete "src/main/resources/www/${h2oRESTApiVersion}/xgboost4j-0.7-jar-with-dependencies.jar"
 }
 
 
@@ -223,11 +211,10 @@ processResources.dependsOn compileHelpFiles
 processResources.dependsOn copyHelpImages
 processResources.dependsOn copyGenModelJarToWebRoot
 processResources.dependsOn copyDeepWaterJarToWebRoot
-processResources.dependsOn copyXGBoostJarToWebRoot
 
 compileAndInstallDocFiles.dependsOn installNpmPackages
 compileAndInstallDocFiles.dependsOn ":h2o-docs:build"
 build.dependsOn compileAndInstallDocFiles
 
 // Hook clean stage
-clean.dependsOn deleteFlowFromWebRoot, cleanUpSmokeTest, deleteGenModelJarFromWebRoot, deleteDeepWaterJarFromWebRoot, deleteXGBoostJarFromWebRoot, deleteDocFiles
+clean.dependsOn deleteFlowFromWebRoot, cleanUpSmokeTest, deleteGenModelJarFromWebRoot, deleteDeepWaterJarFromWebRoot, deleteDocFiles
