+++ py/testdir_multi_jvm/test_rest_api.py
@@ -22,7 +22,7 @@ sys.path.insert(1, '../rest_tests')
 
 clean_up_after = False
 
-algos = ['kmeans', 'deeplearning', 'deepwater', 'drf', 'glm', 'gbm', 'pca', 'naivebayes', 'glrm', 'svd', 'aggregator', 'word2vec', 'stackedensemble']
+algos = ['kmeans', 'deeplearning', 'deepwater', 'drf', 'glm', 'gbm', 'pca', 'naivebayes', 'glrm', 'svd', 'aggregator', 'word2vec', 'stackedensemble', 'xgboost']
 algo_additional_default_params = { 'grep' : { 'regex' : '.*' },
                                    'kmeans' : { 'k' : 2 }
                                  } # additional params to add to the default params
